range-diff: zero out elements in "cost" first
Change the get_correspondences() function so that the zeroing out of
elements in the "cost" array comes first. This changes no behavior,
but makes it easier to demonstrate an existing overflow and segfault
in this code in a subsequent commit.

As can be seen from a careful reading of this code added in in
d9c66f0b5bf (range-diff: first rudimentary implementation,
2018-08-13) (as well as adding some printf-debugging) we'll set all
elements in the in the "n * n" allocated array. That's "n = A+B" where
"A" and "B" are the number of commits in our respective ranges.

So we could, and probably should, just allocate it with
CALLOC_ARRAY(), and skip these two for-loops, but since I'll need them
to quickly demonstrate a segfault let's keep them, "quickly" because
we potentially overflow "int" here, and this is by far the cheapest
set of loops in this function.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>