format-patch tests: extend "check_headers" to test Message-ID
Cover a blindspot in the "format-patch" tests and check what the
Message-ID looks like for corresponding subjects generated with
various command-line options, our behavior differs between cover
letters and regular patches, but nothing was testing for the format
we're emitting.

I have some changes planned for later to make the Message-ID prettier,
but let's first test what we're generating now.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>