Makefile: don't double-check GENERATE_COMPILATION_DATABASE on "yes"
When the "GENERATE_COMPILATION_DATABASE" flag was added in
3821c380687 (Makefile: add support for generating JSON compilation
database, 2020-09-03) positively asserts that you'd like to have a
compilation database generated its implementation was copy/pasted from
the "COMPUTE_HEADER_DEPENDENCIES" flag in a way that didn't make much
sense.

For "COMPUTE_HEADER_DEPENDENCIES" we default to "auto", and want to
probe the compiler to see if we support generating the header
dependency graph, so we need a "dep_check".

But for "GENERATE_COMPILATION_DATABASE" we default to it being
disabled, and the developer has to opt-in by defining it (to
"yes"). There's no point in a pro-active "compdb_check", we can just
add the flag to the $(CC) command-line and have the compiler error out
if it isn't supported.

This removes quite a bit of complexity in the Makefile, and makes the
flow of this logic much more obvious. See [1] for a past mailing list
discussion and discussion of the trade-offs involved here[1].

This changes the semantics of the variable to make "no" mean
"defined", but that should be OK, the "no" was another thing that was
copy/pasted from "COMPUTE_HEADER_DEPENDENCIES". For this flag we've
always documented that it should be set "to yes", let's just change
the semantics here and say "define" instead (which we for most other
such flags).

1. https://lore.kernel.org/git/20210922183311.3766-1-carenas@gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>