git: implement --config-key and --config-env
Implement a way to set config variables that contain the "=" character
on the command-line. This works around an inherent limitation of the
"-c" option, the "-c" option itself can't be changed, we'll always
split on the first "=":

    $ git -c core.name="a.value=with=equals" config core.name
    a.value=with=equals

We can't implement this as "git -c key value" either, since the
"value" might be the git command name, and we already support
value-less "key" to set key to true:

    $ git -c core.name config --type bool core.name
    true

See f9dbb64fadf (config: parse more robust format in
GIT_CONFIG_PARAMETERS, 2021-01-12) for the relevant scaffolding that
we're now making use of, and 1ff21c05ba9 (config: store "git -c"
variables using more robust format, 2021-01-12) for the storage side
of it. As the latter commit notes while the "GIT_CONFIG_PARAMETERS"
format supported this, nothing made use of it, until now.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>