parse-options: add a flag to support "-h" exception
For all of the built-ins except "blame", "ls-remote" and "show-ref" a
"-h" anywhere on the command-line should mean "show the
help".

However, to support the use-case of "-h" being a historically
unfortunate name for another option in "ls-remote" and "show-ref" this
has been guarded with a check for "ctx->total == 1" added in
5ad0d3d5266 (parse-options: allow -h as a short option, 2015-11-17).

Thus "ls-remote" and "show-ref" would make it past the check in
parse_options_step(), and parse their own 'h' option.

The previous interface was very prone to the accidental introduction
of bugs. As seen in the preceding commit the "grep" command
understands "-h", but did not pass the "PARSE_OPT_NO_INTERNAL_HELP"
flag, its "-h" option and help output worked as an emergent effect of
the previous "ctx->total == 1" rule.

Furthermore nothing checked that an API user that didn't pass a
"PARSE_OPT_NO_INTERNAL_HELP" wasn't defining its own 'h' option (which
couldn't be done, as it would trample on "ls-remote" and "show-ref").

To address this let's add a new "PARSE_OPT_INTERNAL_HELP_ON_ONE_ARG"
flag to be used by these special snowflakes. This limits the
"ctx->total == 1" check to them specifically. Internally the flag
implies "PARSE_OPT_NO_INTERNAL_HELP", but as that's an implementation
detail combining them when using the interface is an error.

We then need to add a "PARSE_OPT_REV_PARSE_PARSEOPT" flag to opt-in
the "git rev-parse --parseopt" user, since its option specs may
contain a "h|help", or not, and in both cases we'd like to not have
"PARSE_OPT_NO_INTERNAL_HELP" set.

We'll furthermore make it an error to have handling of any of 'h',
"help" or "help-all" in the "struct options" array when none of the
"PARSE_OPT_NO_INTERNAL_HELP", "PARSE_OPT_INTERNAL_HELP_ON_ONE_ARG" or
"PARSE_OPT_REV_PARSE_PARSEOPT" flags are set. This will help to catch
future logic errors during development.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>