refs: plumb repo into ref stores
In preparation for the next 2 patches that adds (partial) support for
arbitrary repositories to ref iterators, plumb a repository into all ref
stores. There are no changes to program logic.

(The repository is plumbed into the ref stores instead of directly into
the ref iterators themselves, so that existing code that operates on ref
stores do not need to be modified to also handle repositories.)

Signed-off-by: Jonathan Tan <jonathantanmy@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>