fixup! reftable: add a heap-based priority queue for reftable records
originally discussed as part of the -DNDEBUG thread[1], but still likely
to break the build with the parent variable set but not used when assert
gets compiled out.

moving to BUG or DIE still a better option otherwise, but this fixup will
at least address that specific problem (even if the function then is a
NOOP).

[1] https://lore.kernel.org/git/20210902053023.44006-4-carenas@gmail.com/

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>