squash! reftable: reading/writing blocks
Avoid a possible infinite loop in platforms where size_t > long
by using instead the required type and checking for overflow
before doubling.

Increase slightly the initial buffer based on the zlib(1)
recommendation to reduce the likelihood of this code triggering,
and add some error checking to the memory allocation.

[1] https://refspecs.linuxbase.org/LSB_3.0.0/LSB-Core-generic/LSB-Core-generic/zlib-compress2-1.html

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>