COPYING: move GPL text to a "GPL-2.0" file
The preceding commit changed the "COPYING" file to a summary of Git
being licensed under the GPL v2.0, and a description of us carrying
various other code under different licenses.

Let's move the text of the GPL v2.0 to its own file, both to avoid the
wall of text in the "COPYING" summary, and to bring us in line with
both our own "LGPL-2.1" file, and linux.git's "COPYING" file.

As noted in the preceding commit linux.git has not had a copy of the
GPL v2.0 in its top-level "COPYING file since [2].

I think that just a "see GPL-2.0" should be sufficient to point the
reader to it. E.g. emacs.git uses similar language in various places, or:

    License: GNU General Public License version 3 or later (see COPYING)

1. 36268b762c4 (provide a copy of the LGPLv2.1, 2011-05-19)
2. torvalds/linux@bf02d491237e (COPYING: use the new text with points
   to the license files, 2018-03-23)

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>