grep: don't emit stat() permission error twice
When there's a file that "grep" can't stat() it'll try to load it in
grep_source_1() with a call to grep_source_is_binary(), but since
"can't load" is treated the same as "isn't binary" we'll go on to try
to do the reading of the file again in by calling fill_textconv_grep()
below that in the grep_source_1() function.

This isn't pretty, but works around that issue. Perhaps it would be
better to communicate this all the way up with return values, but this
solves the isolated case we care about.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>